perl script to compare Project results with RPAL
using the following command, we can compare the output of this project with that of the RPAL interpreter. We can run these 2 on the same test cases and check what the differences are. Also, a separate folder gets created named results with the test cases that do not match the RPAL output.
Command:

perl difftest.pl -1 "./rpal FILE" -2 "./p2 FILE" -t ./tests/ -r results -v 2